// frontend/src/sections/settings/configuration/nameProperties.generated.ts
/* eslint-disable */
// THIS FILE IS GENERATED by codegen/plugin.js — do not edit by hand.
import type { TypedDocument } from "../../../api/graphql";

export type NamePropertiesVariables = Record<string, never>;

export type NamePropertiesResult = {
  nameProperties: ({
  __typename: "NamePropertyConnector";
} & {
  nodes: Array<{
  id: string;
  remoteEditable: boolean;
  property: {
  id: string;
  key: string;
  name: string;
  allowedValues: string | null;
  valueType: "STRING" | "BOOLEAN" | "INTEGER" | "FLOAT" | "DATE";
};
}>;
});
};

export const NameProperties = {
  query: "query nameProperties {\n  nameProperties {\n    ... on NamePropertyConnector {\n      __typename\n      nodes {\n        id\n        remoteEditable\n        property {\n          id\n          key\n          name\n          allowedValues\n          valueType\n        }\n      }\n    }\n  }\n}",
} as TypedDocument<NamePropertiesResult, NamePropertiesVariables>;

export type ConfigureNamePropertiesVariables = {
  input: Array<{
    id: string;
    key: string;
    propertyId: string;
    name: string;
    valueType: "STRING" | "BOOLEAN" | "INTEGER" | "FLOAT" | "DATE";
    allowedValues?: string | null;
    remoteEditable: boolean;
  }>;
};

export type ConfigureNamePropertiesResult = {
  centralServer: {
  general: {
  configureNameProperties: ({
  __typename: "Success";
} & {
  success: boolean;
});
};
};
};

export const ConfigureNameProperties = {
  query: "mutation configureNameProperties($input: [ConfigureNamePropertyInput!]!) {\n  centralServer {\n    general {\n      configureNameProperties(input: $input) {\n        ... on Success {\n          __typename\n          success\n        }\n      }\n    }\n  }\n}",
} as TypedDocument<ConfigureNamePropertiesResult, ConfigureNamePropertiesVariables>;

export type StorePropertiesVariables = {
  storeId: string;
};

export type StorePropertiesResult = {
  names: ({
  __typename: "NameConnector";
} & {
  nodes: Array<{
  id: string;
  properties: string;
}>;
});
};

export const StoreProperties = {
  query: "query storeProperties($storeId: String!) {\n  names(storeId: $storeId, filter: {isStore: true}, page: {first: 5000}) {\n    ... on NameConnector {\n      __typename\n      nodes {\n        id\n        properties\n      }\n    }\n  }\n}",
} as TypedDocument<StorePropertiesResult, StorePropertiesVariables>;